+++ src/globalStyles.js
@@ -1,6 +1,5 @@
 export default {
   html: { height: '100%' },
   body: { height: '100%' },
-  '#___gatsby': { height: '100%' },
-  '#gatsby-focus-wrapper': { height: '100%' },
+  '#__next': { height: '100%' },
 };
