+++ src/pages/_app.js
@@ -1,5 +1,17 @@
+import { ThemeProvider, CSSReset } from '@chakra-ui/core';
+import { Global } from '@emotion/core';
+import customTheme from '../theme';
+import globalStyles from '../globalStyles';
+import 'typeface-open-sans';
+
 const App = ({ Component, pageProps }) => {
-  return <Component {...pageProps} />;
+  return (
+    <ThemeProvider theme={customTheme}>
+      <CSSReset />
+      <Global styles={globalStyles} />
+      <Component {...pageProps} />
+    </ThemeProvider>
+  );
 };
 
 export default App;
